login añadido

+++ infrastructure/driven-adapters/r2dbc-mysql/src/main/java/co/com/powerup/r2dbc/RolReactiveRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.query.ReactiveQueryByExampleExecutor;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Mono;
 
-public interface RolReactiveRepository extends ReactiveCrudRepository<RolEntity, Integer>, ReactiveQueryByExampleExecutor<RolEntity> {
+public interface RolReactiveRepository extends ReactiveCrudRepository<RolEntity, Long>, ReactiveQueryByExampleExecutor<RolEntity> {
     public Mono<Boolean> existsById(Long id);
 }
